Lib/sched.py
"""A generally useful event scheduler class.

Each instance of this klasa manages its own queue.
No multi-threading jest implied; you are supposed to hack that
yourself, albo use a single instance per application.

Each instance jest parametrized przy two functions, one that jest
supposed to zwróć the current time, one that jest supposed to
implement a delay.  You can implement real-time scheduling by
substituting time oraz sleep z built-in module time, albo you can
implement simulated time by writing your own functions.  This can
also be used to integrate scheduling przy STDWIN events; the delay
function jest allowed to modify the queue.  Time can be expressed as
integers albo floating point numbers, jako long jako it jest consistent.

Events are specified by tuples (time, priority, action, argument, kwargs).
As w UNIX, lower priority numbers mean higher priority; w this
way the queue can be maintained jako a priority queue.  Execution of the
event means calling the action function, dalejing it the argument
sequence w "argument" (remember that w Python, multiple function
arguments are be packed w a sequence) oraz keyword parameters w "kwargs".
The action function may be an instance method so it
has another way to reference private data (besides global variables).
"""

# XXX The timefunc oraz delayfunc should have been defined jako methods
# XXX so you can define new kinds of schedulers using subclassing
# XXX instead of having to define a module albo klasa just to hold
# XXX the global state of your particular time oraz delay functions.

zaimportuj time
zaimportuj heapq
z collections zaimportuj namedtuple
spróbuj:
    zaimportuj threading
wyjąwszy ImportError:
    zaimportuj dummy_threading jako threading
z time zaimportuj monotonic jako _time

__all__ = ["scheduler"]

klasa Event(namedtuple('Event', 'time, priority, action, argument, kwargs')):
    def __eq__(s, o): zwróć (s.time, s.priority) == (o.time, o.priority)
    def __lt__(s, o): zwróć (s.time, s.priority) <  (o.time, o.priority)
    def __le__(s, o): zwróć (s.time, s.priority) <= (o.time, o.priority)
    def __gt__(s, o): zwróć (s.time, s.priority) >  (o.time, o.priority)
    def __ge__(s, o): zwróć (s.time, s.priority) >= (o.time, o.priority)

_sentinel = object()

klasa scheduler:

    def __init__(self, timefunc=_time, delayfunc=time.sleep):
        """Initialize a new instance, dalejing the time oraz delay
        functions"""
        self._queue = []
        self._lock = threading.RLock()
        self.timefunc = timefunc
        self.delayfunc = delayfunc

    def enterabs(self, time, priority, action, argument=(), kwargs=_sentinel):
        """Enter a new event w the queue at an absolute time.

        Returns an ID dla the event which can be used to remove it,
        jeżeli necessary.

        """
        jeżeli kwargs jest _sentinel:
            kwargs = {}
        event = Event(time, priority, action, argument, kwargs)
        przy self._lock:
            heapq.heappush(self._queue, event)
        zwróć event # The ID

    def enter(self, delay, priority, action, argument=(), kwargs=_sentinel):
        """A variant that specifies the time jako a relative time.

        This jest actually the more commonly used interface.

        """
        time = self.timefunc() + delay
        zwróć self.enterabs(time, priority, action, argument, kwargs)

    def cancel(self, event):
        """Remove an event z the queue.

        This must be presented the ID jako returned by enter().
        If the event jest nie w the queue, this podnieśs ValueError.

        """
        przy self._lock:
            self._queue.remove(event)
            heapq.heapify(self._queue)

    def empty(self):
        """Check whether the queue jest empty."""
        przy self._lock:
            zwróć nie self._queue

    def run(self, blocking=Prawda):
        """Execute events until the queue jest empty.
        If blocking jest Nieprawda executes the scheduled events due to
        expire soonest (jeżeli any) oraz then zwróć the deadline of the
        next scheduled call w the scheduler.

        When there jest a positive delay until the first event, the
        delay function jest called oraz the event jest left w the queue;
        otherwise, the event jest removed z the queue oraz executed
        (its action function jest called, dalejing it the argument).  If
        the delay function returns prematurely, it jest simply
        restarted.

        It jest legal dla both the delay function oraz the action
        function to modify the queue albo to podnieś an exception;
        exceptions are nie caught but the scheduler's state remains
        well-defined so run() may be called again.

        A questionable hack jest added to allow other threads to run:
        just after an event jest executed, a delay of 0 jest executed, to
        avoid monopolizing the CPU when other threads are also
        runnable.

        """
        # localize variable access to minimize overhead
        # oraz to improve thread safety
        lock = self._lock
        q = self._queue
        delayfunc = self.delayfunc
        timefunc = self.timefunc
        pop = heapq.heappop
        dopóki Prawda:
            przy lock:
                jeżeli nie q:
                    przerwij
                time, priority, action, argument, kwargs = q[0]
                now = timefunc()
                jeżeli time > now:
                    delay = Prawda
                inaczej:
                    delay = Nieprawda
                    pop(q)
            jeżeli delay:
                jeżeli nie blocking:
                    zwróć time - now
                delayfunc(time - now)
            inaczej:
                action(*argument, **kwargs)
                delayfunc(0)   # Let other threads run

    @property
    def queue(self):
        """An ordered list of upcoming events.

        Events are named tuples przy fields for:
            time, priority, action, arguments, kwargs

        """
        # Use heapq to sort the queue rather than using 'sorted(self._queue)'.
        # With heapq, two events scheduled at the same time will show w
        # the actual order they would be retrieved.
        przy self._lock:
            events = self._queue[:]
        zwróć list(map(heapq.heappop, [events]*len(events)))
